prospect-prediction/requirements.txt
langgraph
langchain
langchain-community
langchain-google-vertexai
google-genai>=0.3.0
langchain-core
chainlit
vertexai
google-cloud-discoveryengine
requests
pandas
scikit-learn
db-dtypes